put face animation to separate file
