+++ firmware/src/oled.h
@@ -6,6 +6,9 @@
 #define TASK_OLED_FREQUENCY 100                             // Hz
 #define TASK_OLED_INTERVAL  (1000 / TASK_OLED_FREQUENCY)   // ms
 
+#define SCREEN_WIDTH        128 // OLED display width, in pixels
+#define SCREEN_HEIGHT       64 // OLED display height, in pixels
+
 typedef enum {
     FRAME_SHIFT_LEFT = 0,
     FRAME_SHIFT_RIGHT
@@ -21,11 +24,5 @@ bool Task_OLED_Setup(void);
 void Task_OLED_Periodic(void);
 void Timeline_Draw_Data_Point(uint16_t dataPoint, uint16_t scale, FRAME_Part_e framePart = FRAME_PART_FULL);
 void Plot_ADC_Dump(uint16_t* data, uint16_t size);
-
-void displayFace(uint8_t left_eye_y_offset, uint8_t right_eye_y_offset, uint8_t mouth_y_offset);
-void blinkEyes(const uint16_t counter, const uint16_t end_count);
-void actuateMouth(const uint16_t counter, const uint16_t end_count);
-void squintEyes(const uint16_t counter, const uint16_t end_count);
-void animateFace();
 
 #endif // OLED_H
